use BaseChart for both other charts
